Base skeleton of the game
First idea was to implement Entity-Component system in the same way it
id done ine [1], but it seems that now GenEvent handler requires you to
implement every event, which would lead to wildcard event handle in
every component.

Current implementation spawn new agent for every component and component
itself is nothing more then data that it holds, like described in [2].
State can be get and set and all the work is done when cycling through
list of systems.

[1] https://github.com/davidgchaves/zelda-cat
[2] http://vasir.net/blog/game-development/how-to-build-entity-component-system-in-javascript